Add utests for restart
